+++ loveship/components/widgets/Badge.tsx
@@ -4,7 +4,7 @@ import * as uui from '@epam/uui';
 import { EpamAdditionalColor, EpamPrimaryColor } from '../types';
 import css from './Badge.module.scss';
 
-const defaultSize = '18';
+const DEFAULT_SIZE = '18';
 
 type BadgeColor = EpamPrimaryColor | EpamAdditionalColor | 'yellow'| 'orange' | 'purple' | 'cyan' | 'mint' | 'white'
 | 'night100' | 'night300' | 'night600' | uui.BadgeProps['color'];
@@ -36,7 +36,7 @@ function applyBadgeMods(mods: BadgeMods) {
     return [
         css['style-' + (mods.shape || 'square')],
         css[`fill-${mods.fill === 'semitransparent' ? 'outline' : (mods.fill || 'solid')}`],
-        css['size-' + (mods.size || defaultSize)],
+        css['size-' + (mods.size || DEFAULT_SIZE)],
         css.root,
     ];
 }
@@ -55,10 +55,17 @@ export const Badge = createSkinComponent<uui.BadgeProps, BadgeProps>(
                 propValueUseInstead: 'outline',
                 condition: () => ['semitransparent'].indexOf(props.fill) !== -1,
             });
+            devLogger.warnAboutDeprecatedPropValue<BadgeProps, 'shape'>({
+                component: 'Badge',
+                propName: 'shape',
+                propValue: props.shape,
+                propValueUseInstead: 'round',
+                condition: () => props.shape === 'square' || !props.shape,
+            });
         }
         return {
             color: props.color || 'sky',
-            size: props.size || defaultSize,
+            size: props.size || DEFAULT_SIZE,
             fill: props.fill === 'semitransparent' ? 'outline' : (props.fill || 'solid'),
         };
     },