+++ changelog.md
@@ -8,6 +8,7 @@
 * [DataTable]: Focus manager.
   * Added ability to change focus on different cells using a keyboard in editable tables.
   * See the example [here](https://uui.epam.com/demo?id=editableTable).
+* [Badge]: in Loveship property `shape: square` was deprecated and will be removed in future release. Please pay attention to the `sguare` value set by default now, and we recommend changing it to `round` in this case.
 
 **What's Fixed**
 * [LazyDataSource]: fixed cascade selection with not flatten search.